requirements.txt
openai
langchain
langchain-community
faiss-cpu
sentence-transformers
python-dotenv
PyMuPDF
transformers